Add README.docx
This file contains the images of the website. It shows the signup and login page initially. After login, it shows the page with all the events. The next one is the event list created by current logged in user and next to that is the events liked by the current logged in user. At last we see the log out functionality and the change in navigation bar due to it.